# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "srvcs-slope",
    "description": "",
    "license": {
      "name": "Apache-2.0",
      "identifier": "Apache-2.0"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/": {
      "get": {
        "tags": [],
        "summary": "`GET /` — service identity (srvcs service standard).",
        "operationId": "index",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Info"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [],
        "summary": "`POST /` — compute the slope of the line through two points.",
        "description": "This service owns the *control flow* but delegates every arithmetic step to\nits dependencies, exactly as specified:\n\n1. `dy = floatsubtract(y2, y1)`;\n2. `dx = floatsubtract(x2, x1)`;\n3. `result = floatdivide(dy, dx)` — `srvcs-floatdivide` returns `422` on a\n   vertical line (`dx == 0`), which is forwarded verbatim.\n\nValidation is not handled here: this service never calls `srvcs-isnumber`\ndirectly. Its dependencies validate their own operands, and any `422` they\nraise (a non-numeric coordinate, or a vertical line) is forwarded.",
        "operationId": "evaluate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EvalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SlopeResponse"
                }
              }
            }
          },
          "422": {
            "description": "a dependency rejected an input, or the line is vertical (forwarded)"
          },
          "500": {
            "description": "a dependency returned a malformed result"
          },
          "503": {
            "description": "a dependency is unavailable"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "EvalRequest": {
        "type": "object",
        "required": [
          "x1",
          "y1",
          "x2",
          "y2"
        ],
        "properties": {
          "x1": {
            "type": "object",
            "description": "The x-coordinate of the first point."
          },
          "x2": {
            "type": "object",
            "description": "The x-coordinate of the second point."
          },
          "y1": {
            "type": "object",
            "description": "The y-coordinate of the first point."
          },
          "y2": {
            "type": "object",
            "description": "The y-coordinate of the second point."
          }
        }
      },
      "Info": {
        "type": "object",
        "required": [
          "service",
          "concern",
          "depends_on"
        ],
        "properties": {
          "concern": {
            "type": "string"
          },
          "depends_on": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "service": {
            "type": "string"
          }
        }
      },
      "SlopeResponse": {
        "type": "object",
        "required": [
          "x1",
          "y1",
          "x2",
          "y2",
          "result"
        ],
        "properties": {
          "result": {
            "type": "number",
            "format": "double"
          },
          "x1": {
            "type": "object"
          },
          "x2": {
            "type": "object"
          },
          "y1": {
            "type": "object"
          },
          "y2": {
            "type": "object"
          }
        }
      }
    }
  }
}